+++ ta4j-core/src/test/java/org/ta4j/core/analysis/CumulativePnLTest.java
@@ -96,4 +96,15 @@ public class CumulativePnLTest extends AbstractIndicatorTest<org.ta4j.core.Indic
         assertNumEquals(0, pnl.getValue(2));
     }
 
+    @Test
+    public void markToMarketCanIgnoreOpenPosition() {
+        var series = new MockBarSeriesBuilder().withNumFactory(numFactory).withData(100, 105, 102).build();
+        var record = new BaseTradingRecord(Trade.buyAt(0, series));
+
+        var pnl = new CumulativePnL(series, record, EquityCurveMode.MARK_TO_MARKET, OpenPositionHandling.IGNORE);
+        assertNumEquals(0, pnl.getValue(0));
+        assertNumEquals(0, pnl.getValue(1));
+        assertNumEquals(0, pnl.getValue(2));
+    }
+
 }